Add actions to publish to gh-pages (#3)
* Add actions to publish to gh-pages (#1)

Actions have been added to publish to the `gh-pages` branch. Each
deployment is made to a different subdirectory, depending on the
circumstances of the deployment.

The `main` branch gets deployed to the subdirectory `latest/`, so that
we always have a "live" version of the current latest development
branch.

Any release candidates get deployed to the subdirectory
`rc-[release version]`, so that we can properly test the release
candidate.

After a release has been tagged, it gets deployed to the subdirectory
`[release version]`. This will be the version used in production by
MetaMask.

* Add `contents: write` permission

The `contents: write` permission has been added to all `gh-pages`
publishing actions. This ensures they have permission to write to the
`gh-pages` branch.

The permission was required on the reusable workflow where the actual
deploy happens, and on the caller workflows.