Add favicons (#8)
Two favicons have been added: an SVG file for newer browsers, and an
"ico" file for older browsers. The "ico" file is a 32x32 png generated
from the SVG using inkscape.

The favicon is supposed to be square, so I modified the fox SVG file to
be square (it was 35x33, I made it 35x35). There is now slightly more
whitespace above and below the fox in the SVG, but this doesn't affect
how it's rendered on the page because we're already setting the icon as
a square using CSS.