Remove unnecessary location test mock
The location mock used in unit tests has been removed. Instead we now
navigate directly in the test by setting `document.location.href`.

jsdom only allows hash-navigations right now, so for this to be
possible the base URL for the environment needed to be set. The base
URL was moved to a separate module so that it could be shared between
the tests and the base Jest config. A dummy version of `vX.Y.Z` was
also added to this base URL so that it more closely matches the real
production URL.