Fix `gh-pages` publishing
The `gh-pages` actions were not triggering because they are triggered
by a tag. The tag was being added by GitHub actions, so the trigger
failed because GitHub actions ignores self-initiated events.

A personal access token belonging to the `metamaskbot` account has been
added to this repository as a secret. This allows us to tag using the
bot's account, ensuring the later triggers work correctly.

Fixes #13