Dynamically lookup the source of a block
When a user chooses to dispute a block, we now dynamically lookup the
source of the block instead of getting the dispute URL from the query
parameters. This is in preparation for an extension update that will
necessitate dropping the `newIssueUrl` query parameter we had
previously been relying on.

This is a temporary measure. A future update will restore the source
for each block. At that time we can revert this change.

This should be comparing in exactly the same way as the extension.
The configuration is fetched dynamically so it may be slightly more
up-to-date than what the extension is using.