Migrate from Jest to Playwright
The Jest tests have been completely replaced by e2e tests written using
Playwright.

Jest is not well suited for testing the behavior of this website. Too
much of the code-under-test was related to DOM/browser behavior that is
not adequately mocked by Jest's `jsdom` environment, and there was no
way to fully reset in between each test to prevent them from
interfering with each other. As a result, most of the functionality was
not being tested.

We should now have nearly complete coverage. The tests have been split
into modules, one for each use-case.

The dependency `http-server` was added to serve the site locally during
development, and in CI.

The README has been updated with some basic instructions, but there is
still much room for improvement there.

The Playwright config was auto-generated by Playwright's setup command,
with minimal changes made. The change from default were:
* Set a `baseURL`
* Uncomment Edge and Google Chrome
* Set the `webServer` config