Fix suspect href initialization (#59)
* Fix suspect href initialization

The blocked href is meant to be included in the text of the page.
Previously it was not being initialized properly because it was being
set to `document.referrer`. But the referrer is not set when the
extension opens this page in practice, and even if it was, it would not
be set to the blocked URL.

Instead it now gets the blocked URL from the query parameters. This
initialization has been moved from the HTML file to `index.ts` to be
alongside the rest of the page initialization as well.

* Add tests