Fix initialization of "detection-repo" span (#58)
The "detection-repo" span lists which repositories/projects are
responsible for this advisory. The idea was that we would update this
dynamically to tell the user who was responsible for the block.

This initialization was moved from the HTML document to `index.ts`,
which is where most of the other page initialization logic is.

The markup was updated to provide a default value, so that the page
looks less broken when we don't know the source.

But most importantly, the way we're checking the source was fixed.
Previously we were relying on `window.location.href` including a string
that would indicate the source, but this is never set in practice.

Instead it now looks at the `newIssueUrl` query parameter, which is
currently how the extension communicates the source of each block.