Update GitHub Actions to match template (#63)
The GitHub Actions configuration has been updated to match the module
template more closely. Some differences remain because this repository
is publishing a website, rather than docs.

The high-level changes are:
* The Node.js version matrix has been updated to include v18 and v19.
This repository is not really a library and it's only used by the
extension, so Node.js v14 support remains unnecessary.
* A GitHub Actions lint has step has been added
* Build, lint, and test now run in parallel
* The actions we use have been updated
* The broken "Require additional reviewer" action has been removed
* The deprecated syntax for setting output has been updated.
* Switched from Ubuntu v20.04 to latest
* The standard release flow is now used, with publishing triggered on
the `main` branch rather than on merge.