Fix release publishing
The recent GitHub Actions changes in #63 broke the last step of the
release process, where the GitHub Release (and git tag) was created. It
failed because we were using a custom token to publish the release, and
it wasn't being properly passed through the `workflow_dispatch` call.

We don't actually need any special token for this step, we can use the
built-in `GITHUB_TOKEN` instead. The workflow has been updated to do
that.