Restore iframe warning and "open in new tab" link (#73)
The "Open in new tab" link that we show on the warning page when it's
rendered in an iframe was accidentally removed in #52. We need to
preserve this button so that users can better understand why their
iframe'd page is blocked, and so they can dispute or bypass the block
if necessary.