Fix GitHub pages release deployment
The deployment to GitHub pages was failing because it was triggered by
a tag, and that tag is now being created inside of a GitHub action
using the `GITHUB_TOKEN`. This prevents the publish from being
triggered, because `GITHUB_TOKEN`-authorized actions never trigger
anything.

The publishing step was moved to `publish-release.yml` instead of being
triggered separately by the tag. This matches the module template.